feat: add admin session check in server actions methods

+++ app/actions/admin/faq-delete.ts
@@ -2,8 +2,12 @@
 
 import { prisma } from "@/app/lib/db";
 import { redirect } from "next/navigation";
+import checkAdmin from "@/app/utils/checkSession";
 
 export async function faqDelete(id: string) {
+    //
+    if (!(await checkAdmin())) return;
+
     await prisma.faq.delete({
         where: {
             id: id,
